requirements.txt
# AI Decision Engine — dependencies
# Install: pip install -r requirements.txt

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0

# API layer
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# OpenAI (optional — runs in simulation mode without API key)
openai>=1.0.0

# Testing
httpx>=0.27.0
